+++ app/src/main/java/io/sikorka/android/ui/wizard/slides/network_selection/NetworkSelectionContract.kt
@@ -5,11 +5,11 @@ import io.sikorka.android.mvp.BaseView
 import io.sikorka.android.mvp.Presenter
 
 interface NetworkSelectionView : BaseView {
-  fun updateNetworkSelection(@Network.Selection network: Long)
+  fun updateNetworkSelection(@Network.Selection network: Int)
 }
 
 
 interface NetworkSelectionPresenter : Presenter<NetworkSelectionView> {
-  fun selectNetwork(@Network.Selection network: Long)
+  fun selectNetwork(@Network.Selection network: Int)
   fun updateSelected()
 }
